[ADD][TEST] Split tests into multiple files
